requirements.txt
streamlit
supabase
fpdf2
audio-recorder-streamlit